fix: cond resolver debugs
